pyproject.toml
[project]
name = "machine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "drf-yasg>=1.21.9",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]
